requirements.txt
Django==5.2.4
python-dotenv==1.0.1
langchain==0.2.0
langchain-community==0.2.0
langchain-google-genai==1.0.0
google-generativeai==0.5.4
